+++ submitr/validators/library_prep_validator.py
@@ -13,9 +13,9 @@ from submitr.validators.decorators import structured_data_validator_finish_hook
 _LIBRARY_SCHEMA_NAME = "Library"
 _ANALYTE_PROPERTY_NAME = "analytes"
 _ASSAY_PROPERTY_NAME = "assay"
-_RNA_SEQ_IDENTIFIER = "bulk_rna_seq"
+_RNA_SEQ = "bulk_rna_seq"
 _LIBRARY_PREP_PROPERTY_NAME = "library_preparation"
-_RNA_SEQ_PROTOCOL_PROPERTY = "rna_seq_protocol"
+_RNA_SEQ_PROTOCOL = "rna_seq_protocol"
 _ANALTYE_SCHEMA_NAME = "Analyte"
 _MOLECULE_PROPERTY_NAME = "molecule"
 _RNA_VALUE_NAME = "RNA"
@@ -48,7 +48,7 @@ def _library_prep_validator(structured_data: StructuredDataSet, **kwargs) -> Non
                             if library_prep := [
                                 lp_item
                                 for lp_item in structured_data.data.get(
-                                    _LIBRARY_PREP_SCHEMA_NAME
+                                    _LIBRARY_PREP_SCHEMA_NAME, []
                                 )
                                 if lp_item.get("submitted_id")
                                 == item.get(_LIBRARY_PREP_PROPERTY_NAME)
@@ -62,12 +62,12 @@ def _library_prep_validator(structured_data: StructuredDataSet, **kwargs) -> Non
                                         f" property {_STRAND_PROPERTY_NAME} is required for RNA"
                                         f" libraries"
                                     )
-                                if assay == _RNA_SEQ_IDENTIFIER and _RNA_SEQ_PROTOCOL_PROPERTY not in library_prep[0]:
+                                if assay == _RNA_SEQ and _RNA_SEQ_PROTOCOL not in library_prep[0]:
                                     # missing rna_seq_protocol for RNA-Seq
                                     structured_data.note_validation_error(
                                         f"{_LIBRARY_SCHEMA_NAME}: {submitted_id}"
                                         f" {_LIBRARY_PREP_SCHEMA_NAME} item {library_prep[0].get('submitted_id')}"
-                                        f" property {_RNA_SEQ_PROTOCOL_PROPERTY} is required for RNA-Seq"
+                                        f" property {_RNA_SEQ_PROTOCOL} is required for RNA-Seq"
                                         f" libraries"
                                     )
                         else:
@@ -94,10 +94,10 @@ def _library_prep_validator(structured_data: StructuredDataSet, **kwargs) -> Non
                                         f" {_LIBRARY_PREP_SCHEMA_NAME} item {library_prep[0].get('submitted_id')}:"
                                         f" property {_STRAND_PROPERTY_NAME} is only for RNA libraries"
                                     )
-                                if _RNA_SEQ_PROTOCOL_PROPERTY in library_prep[0]:
+                                if _RNA_SEQ_PROTOCOL in library_prep[0]:
                                     # DNA analyte with rna_seq_protocol property
                                     structured_data.note_validation_error(
                                         f"{_LIBRARY_SCHEMA_NAME}: {submitted_id}"
                                         f" {_LIBRARY_PREP_SCHEMA_NAME} item {library_prep[0].get('submitted_id')}:"
-                                        f" property {_RNA_SEQ_PROTOCOL_PROPERTY} is only for RNA libraries"
+                                        f" property {_RNA_SEQ_PROTOCOL} is only for RNA libraries"
                                     )